Movie
Screen to list all movies with Name, Year of release, Producer and all
Actors of that movie
Screen to ‘add’ a new movie with the necessary fields with existing
actors and producers. If the user wants to add new ‘Actors’ and
‘Producers’ while creating the movie which are not present in the
database then he should be able to do so while being on the same screen.
‘Listing’ screen should allow user to click on ‘Edit’ and edit all the
details of the movie and save it.